+++ packages/plugin-spoiler/src/editor.tsx
@@ -1,39 +1,38 @@
 import { StatefulPluginEditorProps } from '@edtr-io/core'
-import { styled } from '@edtr-io/ui'
+import { ExpandableBox, EditorThemeProvider } from '@edtr-io/ui'
 import * as React from 'react'
 
-import { SpoilerRenderer } from './renderer'
 import { spoilerState } from '.'
 
-const Input = styled.input({
-  '&:active': {
-    color: '#ffffff'
-  }
-})
-
-export const SpoilerEditor = ({
+export function SpoilerEditor({
   state,
   editable
-}: StatefulPluginEditorProps<typeof spoilerState>) => {
-  const { title } = state
+}: StatefulPluginEditorProps<typeof spoilerState>) {
+  const title = editable ? (
+    <input
+      onChange={e => state.title.set(e.target.value)}
+      value={state.title()}
+      placeholder="Titel eingeben"
+    />
+  ) : state.title() ? (
+    state.title()
+  ) : null
 
   return (
-    <SpoilerRenderer
-      state={state}
-      shown={editable}
-      title={
-        editable ? (
-          <Input
-            onChange={e => {
-              title.set(e.target.value)
-            }}
-            value={title.value}
-            placeholder="Title"
-          />
-        ) : (
-          title.value
-        )
-      }
-    />
+    <EditorThemeProvider
+      theme={{
+        ui: {
+          expandableBox: {
+            toggleBackgroundColor: '#f5f5f5',
+            toggleColor: undefined,
+            containerBorderColor: '#f5f5f5'
+          }
+        }
+      }}
+    >
+      <ExpandableBox title={title} editable={editable}>
+        {state.content.render()}
+      </ExpandableBox>
+    </EditorThemeProvider>
   )
 }