+++ packages/plugin-text/src/factory/editor.tsx
@@ -115,7 +115,7 @@ export const createTextEditor = (
             props.state.set(nextValue)
           }
         }}
-        placeholder={options.placeholder}
+        placeholder={props.editable ? options.placeholder : ''}
         plugins={slatePlugins.current}
         readOnly={!props.focused}
         value={rawState}