+++ scripts/changelog.ts
@@ -154,9 +154,11 @@ async function exec(): Promise<void> {
         '**ui**. Add `ContainerWithConfigButton` component for opening the Settings Overlay.',
         '**plugin-geogebra**. Use ContainerWithConfigButton',
         '**plugin-equations**. Add plugin for aligned equations',
+        '**plugin-hint**. Add plugin for hints',
         '**plugin-image**. Use ContainerWithConfigButton',
         '**plugin-input-exercise**. Add plugin for input exercises',
         '**plugin-sc-mc-exercise**. Add plugin for single-choice & multiple-choice exercises',
+        '**plugin-solution**. Add plugin for solutions',
         '**plugin-video**. Use ContainerWithConfigButton',
         '**plugin-text**. Merge with previous / next text plugin on backspace / delete when at start / end of plugin'
       ],
@@ -169,11 +171,18 @@ async function exec(): Promise<void> {
       ]
     },
     {
-      added: ['**plugin-image**. Maxwidth for images'],
-      fixed: ['**plugin-image**. hide config-overlay in render mode'],
       breakingChanges: [
-        '**plugin-image**. Image-plugin has new required `maxwidth` - state property'
-      ]
+        '**plugin-image**. Image-plugin has new required `maxWidth` - state property',
+        '**ui**. Theming is handled differently. Furthermore, we now use a different (but consistent) naming scheme.'
+      ],
+      added: [
+        '**plugin-h5p**. Add H5p.com plugin',
+        '**plugin-image**. Add MaxWidth for images',
+        '**ui**. Add consistent theming workflow'
+      ],
+      changed: ['**plugin-text**. Hide placeholder in render mode'],
+      fixed: ['**plugin-image**. Hide config-overlay in render mode'],
+      internal: ['**demo**. Log errors only']
     }
   ])